refactor: move domain-related constants and types to config directory
