+++ apps/web/src/services/cost-explorer/pages/admin/AdminAdvancedSettingsAnomalyDetectionConfigurationPage.vue
@@ -17,7 +17,7 @@ import type { DomainConfigGetParameters } from '@/schema/config/domain-config/ap
 import type { DomainConfigSetParameters } from '@/schema/config/domain-config/api-verbs/set';
 import type { DomainConfigModel } from '@/schema/config/domain-config/model';
 
-import { DOMAIN_CONFIG_NAMES } from '@/store/domain/constant';
+import { DOMAIN_CONFIG_NAMES } from '@/store/config/constant';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
 import { usePageEditableStatus } from '@/common/composables/page-editable-status';
